Website/requirements.txt
asgiref==3.5.0
Django==2.2.2
django-autoslug==1.9.8
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2022.1